# qcm-usr-rest/src/main/java/com/emu/apps/qcm/spi/webmvc/rest/UploadRestController.java
package com.emu.apps.qcm.spi.webmvc.rest;

import com.emu.apps.qcm.api.models.Upload;
import com.emu.apps.qcm.domain.ports.UploadBusinessPort;
import com.emu.apps.shared.annotations.Timer;
import com.emu.apps.shared.security.AuthentificationContextHolder;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springdoc.core.converters.models.PageableAsQueryParam;
import org.springframework.context.annotation.Profile;
import org.springframework.data.domain.Pageable;
import org.springframework.data.web.PageableDefault;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.io.IOException;

import static com.emu.apps.qcm.spi.webmvc.rest.ApiRestMappings.PUBLIC_API;
import static com.emu.apps.qcm.spi.webmvc.rest.ApiRestMappings.UPLOADS;
import static org.springframework.data.domain.Sort.Direction.DESC;

@RestController
@Profile("webmvc")
@RequestMapping(value = PUBLIC_API + UPLOADS, produces = MediaType.APPLICATION_JSON_VALUE)
@Tag(name = "Upload")
public class UploadRestController {

    private final UploadBusinessPort uploadBusinessPort;

    public UploadRestController(UploadBusinessPort uploadServicePort) {
        this.uploadBusinessPort = uploadServicePort;
    }

    @ResponseBody
    @PostMapping(value = "/{fileType}")
    public Upload uploadFileByType(@PathVariable("fileType") String fileType,
                                   @RequestParam("file") MultipartFile multipartFile,
                                   @RequestParam(value = "async", required = false) Boolean async) throws IOException {

        return uploadBusinessPort.uploadFile(fileType, multipartFile, async, AuthentificationContextHolder.getUser());
    }

    @GetMapping
    @Timer
    @PageableAsQueryParam
    public Iterable <Upload> getUploads(@Parameter(hidden = true)
                                        @PageableDefault(direction = DESC, sort = {"dateModification"}) Pageable pageable) {
        return uploadBusinessPort.getUploads(pageable, AuthentificationContextHolder.getUser());
    }


    @DeleteMapping(value = "/{uuid}")
    @ResponseBody
    @ResponseStatus(value = HttpStatus.NO_CONTENT)
    public void deleteUploadByUuid(@PathVariable("uuid") String uuid) {
        uploadBusinessPort.deleteUploadByUuid(uuid);
    }

    @GetMapping(value = "/{uuid}")
    @ResponseBody
    public Upload getUploadByUuid(@PathVariable("uuid") String uuid) {
        return uploadBusinessPort.getUploadByUuid(uuid);
    }


}