# swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "version" : "2-oas3",
    "title" : "QCM Rest API",
    "contact" : {
      "name" : "Eric MULLER",
      "url" : "https://qcm-rest-api.herokuapp.com/swagger-ui.html#",
      "email" : "eric.pierre.muller@gmail.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/mit-license.php"
    }
  },
...
  "servers" : [ {
    "url" : "https://qcm-rest-api.herokuapp.com/",
    "description" : "Heroku server"
  }, {
    "url" : "https://qcm-api.webmarks.net/",
    "description" : "webmarks server"
  } ],
  "security" : [ {
    "implicit" : [ "user" ]
  } ],
  "paths" : {
...
    "/qcm/api/v1/upload" : {
      "get" : {
        "tags" : [ "Upload" ],
        "summary" : "Find all uploads by Page",
        "operationId" : "getUploads",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "Results page you want to retrieve (0..N)",
          "required" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page.",
          "required" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully retrieved list",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Upload"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "You are not authorized to view the resource"
          },
          "403" : {
            "description" : "Accessing the resource you were trying to reach is forbidden"
          },
          "404" : {
            "description" : "The resource you were trying to reach is not"
          }
        },
        "deprecated" : false
      }
    },
    "/qcm/api/v1/upload/{fileType}" : {
      "post" : {
        "tags" : [ "Upload" ],
        "summary" : "upload a file",
        "operationId" : "uploadFile_1",
        "parameters" : [ {
          "name" : "fileType",
          "in" : "path",
          "description" : "fileType",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "201" : {
            "description" : "Upload File Created ",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "401" : {
            "description" : "You are not authorized to view the resource"
          },
          "403" : {
            "description" : "Accessing the resource you were trying to reach is forbidden"
          },
          "404" : {
            "description" : "The resource you were trying to reach is not found"
          }
        },
        "deprecated" : false,
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "async" : {
                    "description" : "async",
                    "type" : "boolean"
                  },
                  "file" : {
                    "description" : "file",
                    "type" : "string",
                    "format" : "binary"
                  }
                },
                "required" : [ "file" ]
              }
            }
          }
        }
      }
    },
    "/qcm/api/v1/upload/{id}$" : {
      "delete" : {
        "tags" : [ "Upload" ],
        "summary" : "Delete a upload by ID",
        "operationId" : "deleteUploadById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResponseEntity"
                }
              }
            }
          },
          "204" : {
            "description" : "No Content"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          }
        },
        "deprecated" : false
      }
    },
    "/qcm/api/v1/upload/{id}/import" : {
      "get" : {
        "tags" : [ "Upload" ],
        "summary" : "import a file",
        "operationId" : "uploadFile",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "id",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadDto"
                }
              }
            }
          },
          "202" : {
            "description" : "Import File Launched ",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Upload"
                }
              }
            }
          },
          "401" : {
            "description" : "You are not authorized to view the resource"
          },
          "403" : {
            "description" : "Accessing the resource you were trying to reach is forbidden"
          },
          "404" : {
            "description" : "The resource you were trying to reach is not found"
          }
        },
        "deprecated" : false
      }
    }
  },
  "components" : {
...
    "schemas" : {
...
      "MessageDto" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "title" : "MessageDto"
      },
...
      "ResponseEntity" : {
        "type" : "object",
        "properties" : {
          "body" : {
            "type" : "object"
          },
          "statusCode" : {
            "type" : "string",
            "enum" : [ "CONTINUE", "SWITCHING_PROTOCOLS", "PROCESSING", "CHECKPOINT", "OK", "CREATED", "ACCEPTED", "NON_AUTHORITATIVE_INFORMATION", "NO_CONTENT", "RESET_CONTENT", "PARTIAL_CONTENT", "MULTI_STATUS", "ALREADY_REPORTED", "IM_USED", "MULTIPLE_CHOICES", "MOVED_PERMANENTLY", "FOUND", "MOVED_TEMPORARILY", "SEE_OTHER", "NOT_MODIFIED", "USE_PROXY", "TEMPORARY_REDIRECT", "PERMANENT_REDIRECT", "BAD_REQUEST", "UNAUTHORIZED", "PAYMENT_REQUIRED", "FORBIDDEN", "NOT_FOUND", "METHOD_NOT_ALLOWED", "NOT_ACCEPTABLE", "PROXY_AUTHENTICATION_REQUIRED", "REQUEST_TIMEOUT", "CONFLICT", "GONE", "LENGTH_REQUIRED", "PRECONDITION_FAILED", "PAYLOAD_TOO_LARGE", "REQUEST_ENTITY_TOO_LARGE", "URI_TOO_LONG", "REQUEST_URI_TOO_LONG", "UNSUPPORTED_MEDIA_TYPE", "REQUESTED_RANGE_NOT_SATISFIABLE", "EXPECTATION_FAILED", "I_AM_A_TEAPOT", "INSUFFICIENT_SPACE_ON_RESOURCE", "METHOD_FAILURE", "DESTINATION_LOCKED", "UNPROCESSABLE_ENTITY", "LOCKED", "FAILED_DEPENDENCY", "UPGRADE_REQUIRED", "PRECONDITION_REQUIRED", "TOO_MANY_REQUESTS", "REQUEST_HEADER_FIELDS_TOO_LARGE", "UNAVAILABLE_FOR_LEGAL_REASONS", "INTERNAL_SERVER_ERROR", "NOT_IMPLEMENTED", "BAD_GATEWAY", "SERVICE_UNAVAILABLE", "GATEWAY_TIMEOUT", "HTTP_VERSION_NOT_SUPPORTED", "VARIANT_ALSO_NEGOTIATES", "INSUFFICIENT_STORAGE", "LOOP_DETECTED", "BANDWIDTH_LIMIT_EXCEEDED", "NOT_EXTENDED", "NETWORK_AUTHENTICATION_REQUIRED" ]
          },
          "statusCodeValue" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "title" : "ResponseEntity"
      },
...
      "Upload" : {
        "type" : "object",
        "properties" : {
          "createdBy" : {
            "type" : "string"
          },
          "data" : {
            "type" : "string",
            "format" : "byte"
          },
          "dateCreation" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateModification" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fileName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "lastModifiedBy" : {
            "type" : "string"
          },
          "pathfileName" : {
            "type" : "string"
          },
          "uuid" : {
            "type" : "string"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "title" : "Upload"
      },
      "UploadDto" : {
        "type" : "object",
        "properties" : {
          "dateCreation" : {
            "type" : "string",
            "format" : "date-time"
          },
          "fileName" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "title" : "UploadDto"
      }
    },
    "securitySchemes" : {
      "implicit" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://keycloak.webmarks.net/auth/realms/qcm/protocol/openid-connect/auth",
            "scopes" : { }
          }
        }
      }
    }
  }
}